pyproject.toml
[project]
name = "sidon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Wataru Nakata", email = "wataru9871@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "descript-audio-codec>=1.0.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.5",
    "peft>=0.16.0",
    "pyroomacoustics",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchcodec>=0.7.0",
    "transformers>=4.56.1",
    "wandb>=0.18.7",
    "webdataset>=0.2.100",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
